Fixed batchimport with anomaly injection
